Factor tt::is_array_of_size to its own file
